server/requirements.txt
aiohttp==3.6.2
aiohttp-cors==0.7.0
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
idna==2.9
idna-ssl==1.1.0
multidict==4.7.6
PyYAML==5.3.1
typing-extensions==3.7.4.2
yarl==1.4.2
